Make param string creation more flexible.
Up to now a param name had to either be mapped from str to py var in all
processes or in none. This was not flexibel enough anymore - in
rename_labels the param 'target' is a string, in resample_cube_temporal
it is an object. Therefore param mapping can now be defined per process.
Though a default list is provided.

Signed-off-by: sherrmann <sophie.herrmann@eodc.eu>